requirements.txt
asgiref==3.3.1
astroid==2.4.2
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
whitenoise==5.2.0
wrapt==1.12.1
